+++ capstone/launchZookeeper.sh
@@ -19,12 +19,16 @@ function usage {
 #     CONFIG_FILE=$1
 # fi
 
-
-pids=`ps ax | grep -i 'peer' | grep -v grep | awk '{print $1}'`
+pids=`ps ax | grep -i 'zookeeper' | grep -v grep | awk '{print $1}'`
 for pid in $(eval echo $pids)
 do
-	echo "########### kill -9 $pid #############"
-	kill -9 $pid
+    if [ $pid == $$ ]
+    then
+        echo "Current PID = " $pid
+    else
+       echo "########### kill -9 $pid #############"
+       kill -9 $pid
+    fi
 done
 
 cd $HOME/kafka
